Rename optimizations-server to optimization-server for consistency
