+++ optimization-server/requirements.txt
@@ -2,7 +2,7 @@
 #    uv export --project=pyproject.toml --output-file=requirements.txt
 -e .
 -e ../clients/python-pyo3
-    # via optimizations-server
+    # via optimization-server
 annotated-types==0.7.0 \
     --hash=sha256:1f02e8b43a8fbbc3f3e0d4f0f4bfc8131bcb4eebe8849b8e5c773f3a1c582a53 \
     --hash=sha256:aff07c09a53a08bc8cfccb9c85b05f1aa9a2a6f23728d790723543408344ce89
@@ -183,7 +183,7 @@ clickhouse-connect==0.8.17 \
     --hash=sha256:dffcb1767c684eac1ae9f2fb5c40cc98e23bd3b57f9fb090cd31f4fd2add898b \
     --hash=sha256:f0772cc3de3df5a460c2a5a4ef5b3737da7bfbc93b91e88000cf712f5a7ab5f9 \
     --hash=sha256:f257922b32949f18013bb52e71f50f2eb4667458d73d5782861b46e22081221f
-    # via optimizations-server
+    # via optimization-server
 colorama==0.4.6 ; sys_platform == 'win32' \
     --hash=sha256:08695f5cb7ed6e0531a20572697297273c47b8cae5a63ffc6d6ed5c201be6e44 \
     --hash=sha256:4f1d9991f5acc0ca119f9d443620b77f9d6b33703e51011c16baf57afb285fc6
@@ -217,7 +217,7 @@ execnet==2.1.1 \
 fastapi==0.115.12 \
     --hash=sha256:1e2c2a2646905f9e83d32f04a3f86aff4a286669c6c950ca95b5fd68c2602681 \
     --hash=sha256:e94613d6c05e27be7ffebdd6ea5f388112e5e430c8f7d6494a9d1d88d43e814d
-    # via optimizations-server
+    # via optimization-server
 fastapi-cli==0.0.7 \
     --hash=sha256:02b3b65956f526412515907a0793c9094abd4bfb5457b389f645b0ea6ba3605e \
     --hash=sha256:d549368ff584b2804336c61f192d86ddea080c11255f375959627911944804f4
@@ -269,7 +269,7 @@ httpx==0.28.1 \
     # via
     #   fastapi
     #   openai
-    #   optimizations-server
+    #   optimization-server
     #   tensorzero
 idna==3.10 \
     --hash=sha256:12f65c9b470abda6dc35cf8e63cc574b1c52b11df2c86030af0ac09b01b13ea9 \
@@ -446,7 +446,7 @@ minijinja==2.9.0 \
     --hash=sha256:a72b45788211507a5d08a537cdb258be67e79766185c97f60c1f2812f80b36c9 \
     --hash=sha256:bc1d7403f81a45f708412fbf666c967e4eb7b177becdc248b667820d23523162 \
     --hash=sha256:e987e5421365de59bc9bd621998bab00246378250f06ad67cbe45f160c57b83c
-    # via optimizations-server
+    # via optimization-server
 nodeenv==1.9.1 \
     --hash=sha256:6ec12890a2dab7946721edbfbcd91f3319c6ccc9aec47be7c7e6b7011ee6645f \
     --hash=sha256:ba11c9782d29c27c70ffbdda2d7415098754709be8a7056d79a737cd901155c9
@@ -454,7 +454,7 @@ nodeenv==1.9.1 \
 openai==1.76.2 \
     --hash=sha256:9c1d9ad59e6e3bea7205eedc9ca66eeebae18d47b527e505a2b0d2fb1538e26e \
     --hash=sha256:f430c8b848775907405c6eff54621254c96f6444c593c097e0cc3a9f8fdda96f
-    # via optimizations-server
+    # via optimization-server
 packaging==25.0 \
     --hash=sha256:29572ef2b1f17581046b3a2227d5c611fb25ec70ca1ba8554b24b0e69331a484 \
     --hash=sha256:d443872c98d677bf60f6a1f2f8c1cb748e8fe762d2bf9d3148b5599295b0fc4f
@@ -473,7 +473,7 @@ pydantic==2.11.4 \
     # via
     #   fastapi
     #   openai
-    #   optimizations-server
+    #   optimization-server
 pydantic-core==2.33.2 \
     --hash=sha256:0069c9acc3f3981b9ff4cdfaf088e98d83440a4c7ea1bc07460af3d4dc22e72d \
     --hash=sha256:031c57d67ca86902726e0fae2214ce6770bbe2f710dc33063187a68744a5ecac \
@@ -760,7 +760,7 @@ tiktoken==0.9.0 \
     --hash=sha256:e88f121c1c22b726649ce67c089b90ddda8b9662545a8aeb03cfef15967ddd03 \
     --hash=sha256:f0968d5beeafbca2a72c595e8385a1a1f8af58feaebb02b227229b69ca5357fd \
     --hash=sha256:f32cc56168eac4851109e9b5d327637f15fd662aa30dd79f964b7c39fbadd26e
-    # via optimizations-server
+    # via optimization-server
 tomli==2.2.1 ; python_full_version < '3.11' \
     --hash=sha256:023aa114dd824ade0100497eb2318602af309e5a55595f76b626d6d9f3b7b0a6 \
     --hash=sha256:02abe224de6ae62c19f090f68da4e27b10af2b93213d36cf44e6e1c5abd19fdd \
@@ -810,7 +810,7 @@ typing-extensions==4.13.2 \
     #   anyio
     #   fastapi
     #   openai
-    #   optimizations-server
+    #   optimization-server
     #   pydantic
     #   pydantic-core
     #   pyright
@@ -852,7 +852,7 @@ uuid-utils==0.10.0 \
     --hash=sha256:dbefc2b9113f9dfe56bdae58301a2b3c53792221410d422826f3d1e3e6555fe7 \
     --hash=sha256:ffc49c33edf87d1ec8112a9b43e4cf55326877716f929c165a2cc307d31c73d5
     # via
-    #   optimizations-server
+    #   optimization-server
     #   tensorzero
 uvicorn==0.34.2 \
     --hash=sha256:0e929828f6186353a80b58ea719861d2629d766293b6d19baf086ba31d4f3328 \
