+++ optimization-server/src/optimizations_server/constants.py
