+++ optimization-server/src/optimizations_server/rendering.py
