+++ optimization-server/src/optimizations_server/sft/openai_sft.py
@@ -9,7 +9,7 @@ from openai import AsyncOpenAI
 from openai.types.chat import ChatCompletionMessageParam
 from openai.types.fine_tuning import FineTuningJob
 from tensorzero import AsyncTensorZeroGateway
-from tensorzero.internal_optimizations_server_types import Sample
+from tensorzero.internal_optimization_server_types import Sample
 from typing_extensions import TypedDict
 
 from ..rendering import get_template_env
