+++ ui/Dockerfile
@@ -69,11 +69,13 @@ WORKDIR /app
 
 RUN pnpm run build
 
+# ========== python-build-env ==========
+
 FROM rust:latest AS python-build-env
 RUN apt-get update && apt-get install -y python3
 COPY --from=ghcr.io/astral-sh/uv:0.6.6 /uv /uvx /bin/
 COPY . /build
-WORKDIR /build/optimizations-server
+WORKDIR /build/optimization-server
 # TODO - figure out how to cache the rust build here
 # UV_PYTHON_DOWNLOADS=never
 # https://hynek.me/articles/docker-uv/
@@ -94,9 +96,9 @@ COPY ./ui/entrypoint.sh /app/entrypoint.sh
 
 RUN chmod +x /app/entrypoint.sh
 
-COPY ./optimizations-server /build/optimizations-server
+COPY ./optimization-server /build/optimization-server
 COPY --from=python-build-env /build/clients/python-pyo3 /build/clients/python-pyo3
-COPY --from=python-build-env /build/optimizations-server/.venv  /build/optimizations-server/.venv
+COPY --from=python-build-env /build/optimization-server/.venv  /build/optimization-server/.venv
 # Workaround for uv issue
 #COPY ./clients/python-pyo3 /clients/python-pyo3
