+++ ui/entrypoint.sh
@@ -34,12 +34,12 @@ cd /app
 pnpm run start &
 
 
-source /build/optimizations-server/.venv/bin/activate
+source /build/optimization-server/.venv/bin/activate
 
 # TODO: use 'uv run' once this issue is fixed: https://github.com/astral-sh/uv/issues/9191
 #RUST_LOG=trace uv run --verbose --frozen --no-dev fastapi run --port 7001 src/ &
 
-uv run uvicorn --app-dir /build/optimizations-server --port 7001 optimizations_server.main:app --log-level warning &
+uv run uvicorn --app-dir /build/optimization-server --port 7001 optimization_server.main:app --log-level warning &
 
 # Wait for any process to exit
 wait -n
